Fix warning by -Wpedantic option
In previous implementaion, we use GNU extension to get the label
address, but it doesn't work when appling -Wpedantic which is added by
Freedom-E-SDK CI testing.

There are two ways to fix that. The first one is using diagnostic
pragmas to ignore the pedantic option

"#pragma GCC diagnostic ignored "-Wpedantic""

The second one is changing the implementation by using function address
instead of label address. This is what we to do in this patch.

Signed-off-by: Zong Li <zong.li@sifive.com>